api/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0
httpx>=0.25.1